fix padding on inital loading window
