+++ frontend/src/components/AppLoadingSpinner/styles.scss
@@ -9,4 +9,6 @@
   border-radius: 8px;
   padding: 1.25em;
   overflow: hidden;
+  max-width: 1200px;
+  margin: auto;
 }
