+++ frontend/src/index.scss
@@ -12,6 +12,7 @@ body {
 html {
   overflow-x: hidden;
   margin-right: calc(-1 * (100vw - 100%));
+  scroll-behavior: smooth;
 }
 
 // Background is separate element to allow for multiple layered blurs
